requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
Flask-Bcrypt==1.0.1
PyMySQL==1.1.0
python-dotenv==1.0.0
Pillow==10.2.0
openpyxl==3.1.2
